+++ plugins/woocommerce-admin/storybook/webpack.config.js
@@ -44,28 +44,28 @@ module.exports = ( storybookConfig ) => {
 
 	storybookConfig.plugins.push(
 		...wcAdminWebpackConfig.plugins,
-		new CopyWebpackPlugin( [
-			{
-				from: path.resolve( __dirname, 'wordpress/css' ),
-				to: 'wordpress/css',
-			},
-			{
-				from: path.resolve(
-					__dirname,
-					`../packages/components/build-style/*.css`
-				),
-				to: `./component-css`,
-				flatten: true,
-			},
-			{
-				from: path.resolve(
-					__dirname,
-					`../packages/experimental/build-style/*.css`
-				),
-				to: `./experimental-css`,
-				flatten: true,
-			},
-		] )
+		new CopyWebpackPlugin( {
+			patterns: [
+				{
+					from: path.resolve( __dirname, 'wordpress/css' ),
+					to: 'wordpress/css/[name][ext]',
+				},
+				{
+					from: path.resolve(
+						__dirname,
+						`../packages/components/build-style/*.css`
+					),
+					to: `./component-css/[name][ext]`,
+				},
+				{
+					from: path.resolve(
+						__dirname,
+						`../packages/experimental/build-style/*.css`
+					),
+					to: `./experimental-css/[name][ext]`,
+				},
+			],
+		} )
 	);
 
 	return storybookConfig;